+++ CHANGES.md
@@ -1,4 +1,6 @@
 # CHANGES
 
+- Cleanup: Removed iOS module
+- Feature: Ported LoopMediaPlayer class
 - Feature: Ported react-native-sound
 - Feature: Init